+++ Plugins/Multiplayer/Source/NexusMultiplayer/Public/NMultiplayerLibrary.h
@@ -4,6 +4,7 @@
 #pragma once
 
 #include "CoreMinimal.h"
+#include "NMultiplayerUtils.h"
 #include "UObject/ObjectMacros.h"
 #include "Kismet/BlueprintFunctionLibrary.h"
 #include "NMultiplayerLibrary.generated.h"
@@ -58,7 +59,10 @@ public:
 	 */
 	UFUNCTION(BlueprintCallable, BlueprintPure = false, DisplayName = "? Has Remote Players", Category = "NEXUS|Multiplayer",
 		meta = (WorldContext = "WorldContextObject", ExpandBoolAsExecs="ReturnValue", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#has-remote-players"))
-	static bool HasRemotePlayersExec(UObject* WorldContextObject);
+	static bool HasRemotePlayersExec(UObject* WorldContextObject)
+	{
+		return HasRemotePlayers(WorldContextObject);
+	};
 
 	/**
 	 * Does the current world have locally connected clients only?
@@ -68,7 +72,10 @@ public:
 	 */	
 	UFUNCTION(BlueprintCallable, BlueprintPure = false, DisplayName = "Has Local Players", Category = "NEXUS|Multiplayer",
 		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#has-local-players-only"))
-	static bool HasLocalPlayersOnly(UObject* WorldContextObject);
+	static bool HasLocalPlayersOnly(UObject* WorldContextObject)
+	{
+		return !HasRemotePlayers(WorldContextObject);
+	};
 
 	/**
 	 * Does the current world have locally connected clients only?
@@ -78,7 +85,10 @@ public:
 	 */		
 	UFUNCTION(BlueprintCallable, BlueprintPure = false, DisplayName = "? Has Local Players", Category = "NEXUS|Multiplayer",
 		meta = (WorldContext = "WorldContextObject", ExpandBoolAsExecs="ReturnValue", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#has-local-players-only"))
-	static bool HasLocalPlayersOnlyExec(UObject* WorldContextObject);
+	static bool HasLocalPlayersOnlyExec(UObject* WorldContextObject)
+	{
+		return HasLocalPlayersOnly(WorldContextObject);
+	};
 
 	/**
 	 * Does the current callstack have GameState authority?
@@ -88,7 +98,14 @@ public:
 	 */
 	UFUNCTION(BlueprintCallable, BlueprintPure = false, DisplayName = "Has GameState Authority", Category = "NEXUS|Multiplayer",
 		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#has-gamestate-authority"))
-	static bool HasGameStateAuthority(UObject* WorldContextObject);
+	static bool HasGameStateAuthority(UObject* WorldContextObject)
+	{
+		if (const UWorld* World = N_GET_WORLD_FROM_CONTEXT(WorldContextObject))
+		{
+			return FNMultiplayerUtils::HasGameStateAuthority(World);
+		}
+		return true;
+	}
 
 	/**
 	 * Does the current callstack have GameState authority?
@@ -98,7 +115,10 @@ public:
 	 */
 	UFUNCTION(BlueprintCallable, DisplayName = "? Has GameState Authority", Category = "NEXUS|Multiplayer",
 		meta = (WorldContext = "WorldContextObject", ExpandBoolAsExecs="ReturnValue", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#has-gamestate-authority"))
-	static bool HasGameStateAuthorityExec(UObject* WorldContextObject);
+	static bool HasGameStateAuthorityExec(UObject* WorldContextObject)
+	{
+		return HasGameStateAuthority(WorldContextObject);
+	}
 
 	/**
 	 * Does the current callstack have World authority?
@@ -108,7 +128,14 @@ public:
 	 */	
 	UFUNCTION(BlueprintCallable, BlueprintPure = false, DisplayName = "Has World Authority", Category = "NEXUS|Multiplayer",
 		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#has-world-authority"))
-	static bool HasWorldAuthority(UObject* WorldContextObject);
+	static bool HasWorldAuthority(UObject* WorldContextObject)
+	{
+		if (const UWorld* World = N_GET_WORLD_FROM_CONTEXT(WorldContextObject))
+		{
+			return FNMultiplayerUtils::HasWorldAuthority(World);
+		}
+		return true;
+	};
 
 	/**
 	 * Does the current callstack have World authority?
@@ -118,7 +145,72 @@ public:
 	 */	
 	UFUNCTION(BlueprintCallable, DisplayName = "? Has World Authority", Category = "NEXUS|Multiplayer",
 		meta = (WorldContext = "WorldContextObject", ExpandBoolAsExecs="ReturnValue", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#has-world-authority"))
-	static bool HasWorldAuthorityExec(UObject* WorldContextObject);
+	static bool HasWorldAuthorityExec(UObject* WorldContextObject)
+	{
+		return HasWorldAuthority(WorldContextObject);
+	};
+
+	/**
+	 * Get a player's unique identifier from the APlayerController.
+	 * @param PlayerController The target APlayerController to use when querying for the player identification number.
+	 * @return The player's identifier.
+	 */
+	UFUNCTION(BlueprintCallable, DisplayName = "Get PlayerIdentifier", Category = "NEXUS|Multiplayer",
+			meta = (DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#get-playeridentifier"))
+	static int32 GetPlayerIdentifier(const APlayerController* PlayerController)
+	{
+		return PlayerController->GetPlayerState<APlayerState>()->GetPlayerId();
+	}
+
+	/**
+	 * Get the first player's unique identifier.
+	 * @param WorldContextObject An object to get the UWorld from.
+	 * @return The player's identifier.
+	 */
+	UFUNCTION(BlueprintCallable, DisplayName = "Get First PlayerIdentifier", Category = "NEXUS|Multiplayer",
+		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#get-first-playeridentifier"))
+	static int32 GetFirstPlayerIdentifier(UObject* WorldContextObject)
+	{
+		if (const UWorld* World = N_GET_WORLD_FROM_CONTEXT(WorldContextObject))
+		{
+			return FNMultiplayerUtils::GetFirstPlayerIdentifier(World);
+		}
+		return 0;
+	}
+
+	/**
+	 * Get the APawn for the given player's unique identifier.
+	 * @param WorldContextObject An object to get the UWorld from.
+	 * @param PlayerIdentifier The target identifier to query for.
+	 * @return If found, APawn, or nullptr.
+	 */	
+	UFUNCTION(BlueprintCallable, DisplayName = "Get Pawn From PlayerIdentifier", Category = "NEXUS|Multiplayer",
+		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#get-pawn-from-playeridentifier"))
+	static APawn* GetPawnFromPlayerIdentifier(UObject* WorldContextObject, const int32 PlayerIdentifier)
+	{
+		if (const UWorld* World = N_GET_WORLD_FROM_CONTEXT(WorldContextObject))
+		{
+			return FNMultiplayerUtils::GetPawnFromPlayerIdentifier(World, PlayerIdentifier);
+		}
+		return nullptr;
+	}
+
+	/**
+	 * Get the AActor for the given player's unique identifier.
+	 * @param WorldContextObject An object to get the UWorld from.
+	 * @param PlayerIdentifier The target identifier to query for.
+	 * @return If found, AActor, or nullptr.
+	 */
+	UFUNCTION(BlueprintCallable, DisplayName = "Get PlayerController From PlayerIdentifier", Category = "NEXUS|Multiplayer",
+		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#get-playercontroller-from-playeridentifier"))
+	static AActor* GetPlayerControllerFromPlayerIdentifier(UObject* WorldContextObject, const int32 PlayerIdentifier)
+	{
+		if (const UWorld* World = N_GET_WORLD_FROM_CONTEXT(WorldContextObject))
+		{
+			return FNMultiplayerUtils::GetPlayerControllerFromPlayerIdentifier(World, PlayerIdentifier);
+		}
+		return nullptr;
+	}
 
 	/**
 	 * An explicit check that the network mode of the world is not NM_Client, thus either a listen server (w/ client) or a dedicated server.	 
@@ -127,7 +219,11 @@ public:
 	 */	
 	UFUNCTION(BlueprintCallable, DisplayName = "Is Server", Category = "NEXUS|Multiplayer",
 		meta = (WorldContext = "WorldContextObject", ExpandBoolAsExecs="ReturnValue", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#is-server"))
-	static bool IsServer(UObject* WorldContextObject);
+	static bool IsServer(UObject* WorldContextObject)
+	{
+		const UWorld* World = N_GET_WORLD_FROM_CONTEXT(WorldContextObject);
+		return World ? (World->GetNetMode() != NM_Client) : false;
+	};
 
 	/**
 	 * An explicit check that the network mode of the world is not NM_Client, thus either a listen server (w/ client) or a dedicated server.	 
@@ -136,7 +232,10 @@ public:
 	 */	
 	UFUNCTION(BlueprintCallable, DisplayName = "? Is Server", Category = "NEXUS|Multiplayer",
 		meta = (WorldContext = "WorldContextObject", ExpandBoolAsExecs="ReturnValue", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#is-server"))
-	static bool IsServerExec(UObject* WorldContextObject);
+	static bool IsServerExec(UObject* WorldContextObject)
+	{
+		return IsServer(WorldContextObject);
+	};
 
 	/**
 	 * Is the current session created from the MultiplayerTest editor command?
@@ -144,7 +243,10 @@ public:
 	 */
 	UFUNCTION(BlueprintCallable, BlueprintPure = false, DisplayName = "Is Multiplayer Test", Category = "NEXUS|Multiplayer",
 		meta=(DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#is-multiplayer-test"))
-	static bool IsMultiplayerTest();
+	static bool IsMultiplayerTest()
+	{
+		return FNMultiplayerUtils::IsMultiplayerTest();
+	};
 
 	/**
 	* Is the current session created from the MultiplayerTest editor command?
@@ -152,5 +254,8 @@ public:
 	*/
 	UFUNCTION(BlueprintCallable, BlueprintPure = false, DisplayName = "? Is Multiplayer Test", Category = "NEXUS|Multiplayer",
 		meta = (ExpandBoolAsExecs="ReturnValue", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#is-multiplayer-test"))
-	static bool IsMultiplayerTestExec();
+	static bool IsMultiplayerTestExec()
+	{
+		return FNMultiplayerUtils::IsMultiplayerTest();
+	}
 };